refactor: export UI component from index.tsx
