+++ src/components/Spaces/AddSpace.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, useRef } from "react";
-import Modal from "../UI/Modal";
+import { Modal } from "../UI/";
 interface AddSpaceProps {
   onAddNewSpace: (
     e: React.MouseEvent<HTMLButtonElement, MouseEvent>,
