+++ src/components/Tabs/TabCard.tsx
@@ -3,9 +3,8 @@ import { useState, useEffect, useRef } from "react";
 import { Tab, Direction } from "../../types";
 import MoveToSpace from "./MoveToSpaceBtn";
 import CloseBtn from "./CloseBtn";
-import Tooltip from "../UI/Tooltip";
+import { Tooltip, Dropdown } from "../UI";
 import PinBtn from "./PinBtn";
-import Dropdown from "../UI/Dropdown";
 import KebabBtn from "./KebabBtn";
 import { cn } from "../../utils";
 import TabOrderBtn from "./TabOrderBtn";
