+++ src/components/UI/index.tsx
@@ -1,0 +1,6 @@
+export { default as Dropdown } from "./Dropdown";
+export { default as IconButton } from "./IconButton";
+export { default as Tooltip } from "./Tooltip";
+export { default as Loader } from "./Loader";
+export { default as Modal } from "./Modal";
+export { default as Heading } from "./Heading";
