+++ src/pages/Home/index.tsx
@@ -22,7 +22,7 @@ import ToggleViewBtn from "./ToggleViewBtn";
 import useWindowId from "../../hooks/useWindowId";
 import useLogin from "../../hooks/useLogin";
 import { Tab, Space, SpaceDoc, Direction } from "../../types";
-import Loader from "../../components/UI/Loader";
+import { Loader } from "../../components/UI";
 import { cn } from "../../utils";
 interface Response {
   success: boolean;
